Compilation flags, let's be pedantic

+++ src/main.c
@@ -269,8 +269,8 @@ int main(void)
     printf("\tCPU_MEMORY_ALLOCATED in kilobytes: %llu\n", CPU_MEMORY_ALLOCATED_IN_BYTES / 1024);
 
     // Final check
-    const int MEMORY_LEAK_THRESHOLD = 256;
-    ASSERT(CPU_MEMORY_ALLOCATED_IN_BYTES <= MEMORY_LEAK_THRESHOLD, "\tERROR: Memory leak detected! with a threshold of %d bytes", MEMORY_LEAK_THRESHOLD);
+    const unsigned long long MEMORY_LEAK_THRESHOLD = 256;
+    ASSERT(CPU_MEMORY_ALLOCATED_IN_BYTES <= MEMORY_LEAK_THRESHOLD, "\tERROR: Memory leak detected! with a threshold of %lld bytes", MEMORY_LEAK_THRESHOLD);
 
     return 0;
 }
